split search in Landing & Result
