+++ src/features/search/pages/Search/Search.tsx
@@ -1,92 +1,18 @@
-import { SearchClient } from 'algoliasearch'
-import React, { useCallback } from 'react'
-import { Configure, InstantSearch } from 'react-instantsearch-core'
-import { StatusBar } from 'react-native'
-import AlgoliaSearchInsights from 'search-insights'
-import { v4 as uuidv4 } from 'uuid'
+import React from 'react'
 
-import { BodySearch } from 'features/search/components/BodySearch/BodySearch'
-import { SearchHeader } from 'features/search/components/SearchHeader/SearchHeader'
-import { SearchSuggestions } from 'features/search/components/SearchSuggestions/SearchSuggestions'
 import { useSearch } from 'features/search/context/SearchWrapper'
-import { useSearchHistory } from 'features/search/helpers/useSearchHistory/useSearchHistory'
-import { useSync } from 'features/search/helpers/useSync/useSync'
-import { client } from 'libs/algolia/fetchAlgolia/clients'
-import { env } from 'libs/environment'
-import { useNetInfoContext } from 'libs/network/NetInfoWrapper'
-import { OfflinePage } from 'libs/network/OfflinePage'
-import { Form } from 'ui/components/Form'
-import { Spacer } from 'ui/theme'
-
-const searchInputID = uuidv4()
-const searchClient: SearchClient = {
-  ...client,
-  search(requests) {
-    if (requests.every(({ params }) => !params?.query)) {
-      return Promise.resolve({
-        results: requests.map(() => ({
-          hits: [],
-          nbHits: 0,
-          page: 0,
-          nbPages: 0,
-          hitsPerPage: 0,
-          processingTimeMS: 0,
-          exhaustiveNbHits: false,
-          query: '',
-          params: '',
-        })),
-      })
-    }
-    return client.search(requests)
-  },
-}
-const suggestionsIndex = env.ALGOLIA_SUGGESTIONS_INDEX_NAME
+import { SearchLanding } from 'features/search/pages/SearchLanding/SearchLanding'
+import { SearchResults } from 'features/search/pages/SearchResults/SearchResults'
+import { SearchView } from 'features/search/types'
 
 export function Search() {
-  useSync()
-  const netInfo = useNetInfoContext()
-  const { searchState, isFocusOnSuggestions } = useSearch()
-  const { view } = searchState
-  const { setQueryHistory, queryHistory, addToHistory, removeFromHistory, filteredHistory } =
-    useSearchHistory()
+  const {
+    searchState: { view },
+  } = useSearch()
 
-  const setQueryHistoryMemoized = useCallback(
-    (query: string) => setQueryHistory(query),
-    [setQueryHistory]
-  )
-
-  if (!netInfo.isConnected) {
-    return <OfflinePage />
+  if (view === SearchView.Landing) {
+    return <SearchLanding />
+  } else {
+    return <SearchResults />
   }
-
-  return (
-    <React.Fragment>
-      <StatusBar barStyle="dark-content" />
-      <Form.Flex>
-        <InstantSearch
-          searchClient={searchClient}
-          indexName={suggestionsIndex}
-          insights={{ insightsClient: AlgoliaSearchInsights }}>
-          <Configure hitsPerPage={5} clickAnalytics />
-          <SearchHeader
-            searchInputID={searchInputID}
-            searchView={view}
-            addSearchHistory={addToHistory}
-            searchInHistory={setQueryHistoryMemoized}
-          />
-          <Spacer.Column numberOfSpaces={2} />
-          {isFocusOnSuggestions ? (
-            <SearchSuggestions
-              queryHistory={queryHistory}
-              addToHistory={addToHistory}
-              removeFromHistory={removeFromHistory}
-              filteredHistory={filteredHistory}
-            />
-          ) : (
-            <BodySearch view={view} />
-          )}
-        </InstantSearch>
-      </Form.Flex>
-    </React.Fragment>
-  )
 }
